Move global js to material
